minor tweaks and cleanup

+++ dotnet/Program.cs
@@ -44,21 +44,21 @@ class Program
 
             await client.connect();
             if(!client.connected() ) {
-                Console.WriteLine("Client connection error: " + client.connectionerror());
+                client.info("Client connection error: " + client.connectionerror());
                 return;
             }
-            Console.WriteLine("Client connection success: " + client.connected());
+            client.info("Client connection success: " + client.connected());
 
             var eventid = client.on_client_event((eventObj) => {
-                Console.WriteLine("Client event " + eventObj.evt + ": " + eventObj.reason);
+                client.info("Client event " + eventObj.evt + ": " + eventObj.reason);
             });
-            Console.WriteLine("Client event id: " + eventid);
+            client.info("Client event id: " + eventid);
 
             var (jwt, error, success) = await client.Signin();
-            Console.WriteLine("Signin JWT: " + jwt);
+            client.info("Signin JWT: " + jwt);
 
 
-            Console.WriteLine("Remove Client event: " + eventid);
+            client.info("Remove Client event: " + eventid);
             client.off_client_event(eventid);
 
             var files = new string[] { "testfile.csv" };
@@ -69,19 +69,24 @@ class Program
 
             var workitem = new Workitem { name = "test from dotnet 1", payload = "{\"name\": \"test from dotnet 1\"}" };
             var push_workitem_result = await client.PushWorkitem("rustqueue", workitem, files);
-            Console.WriteLine("PushWorkitem: ", push_workitem_result);
+            client.info("PushWorkitem: ", push_workitem_result);
 
             workitem = await client.PopWorkitem("rustqueue");
-            Console.WriteLine("PopWorkitem: ", workitem);
-            workitem.state = "successful";
-            workitem = await client.UpdateWorkitem(workitem, new string[] { });
-            Console.WriteLine("UpdateWorkitem: ", workitem);
-
-            await client.DeleteWorkitem(workitem.id);
+            if(workitem != null) {
+                client.info("PopWorkitem: ", workitem);
+                workitem.state = "successful";
+                workitem = await client.UpdateWorkitem(workitem, new string[] { });
+                if (workitem != null) {
+                    client.info("UpdateWorkitem: ", workitem);
+                    await client.DeleteWorkitem(workitem.id);
+                }                
+            } else {
+                client.info("No workitem to update");
+            }
 
 
             string results = await client.Query("entities", "{}", "{\"name\": 1}");
-            Console.WriteLine("results: " + results);
+            client.info("results: " + results);
 
             for(var y = 0; y < 5; y++) {
                 var promises = new List<Task<string>>();
@@ -89,16 +94,16 @@ class Program
                     promises.Add(client.Query("entities", "{}", "{\"name\": 1}"));
                 }
                 var result = await Task.WhenAll(promises);
-                Console.WriteLine("results: " + result.Length);
+                client.info("results: " + result.Length);
             }
 
             // // System.Threading.Thread.Sleep(120000);
 
             var aggregate_results = await client.Aggregate("entities", "[]");
-            Console.WriteLine("aggregate results: " + aggregate_results);
+            client.info("aggregate results: " + aggregate_results);
 
             var insert_one_result = await client.InsertOne("entities", "{\"name\": \"test from dotnet\", \"_type\": \"test\"}");
-            Console.WriteLine("insert one result: " + insert_one_result);
+            client.info("insert one result: " + insert_one_result);
 
             dynamic? item = JsonSerializer.Deserialize<ExpandoObject>(insert_one_result, new JsonSerializerOptions { IncludeFields = true });
             if(item == null) throw new Exception("Failed to deserialize insert_one_result");
@@ -110,13 +115,13 @@ class Program
             insert_one_result = JsonSerializer.Serialize(item);
 
             var update_one_result = await client.UpdateOne("entities", insert_one_result);
-            Console.WriteLine("update one result: " + update_one_result);
+            client.info("update one result: " + update_one_result);
 
             var delete_one_result = await client.DeleteOne("entities", _id);
-            Console.WriteLine("delete one result: " + delete_one_result);
+            client.info("delete one result: " + delete_one_result);
 
             var insert_or_update_one_result2 = await client.InsertOne("entities", "{\"name\": \"test insert or update from dotnet\", \"_type\": \"test\"}");
-            Console.WriteLine("insert one result: " + insert_or_update_one_result2);
+            client.info("insert one result: " + insert_or_update_one_result2);
             dynamic? item2 = JsonSerializer.Deserialize<ExpandoObject>(insert_or_update_one_result2, new JsonSerializerOptions { IncludeFields = true });
             if(item2 == null) throw new Exception("Failed to deserialize insert_one_result");
             item2.name = "test insert or update from dotnet updated";
@@ -128,11 +133,11 @@ class Program
             if(string.IsNullOrEmpty(_id2)) throw new Exception("Failed to get _id from insert_one_result");
 
             var delete_many_by_ids_result = await client.DeleteMany("entities", ids: new string[] { _id2 });
-            Console.WriteLine("delete many by ids result: " + delete_many_by_ids_result);
+            client.info("delete many by ids result: " + delete_many_by_ids_result);
 
             await client.InsertOne("entities", "{\"name\": \"test delete many from dotnet\", \"_type\": \"test\"}");
             var delete_many_by_query_result = await client.DeleteMany("entities", query: "{\"name\": \"test delete many from dotnet\"}");
-            Console.WriteLine("delete many by query result: " + delete_many_by_query_result);
+            client.info("delete many by query result: " + delete_many_by_query_result);
 
             await client.download("fs.files", "65a3aaf66d52b8c15131aebd", folder: "", filename: "");
 
@@ -142,14 +147,14 @@ class Program
                 filepath = "../testfile.csv";
             }
             var upload_response = await client.upload(filepath, "dotnet-test.csv", "", "", "fs.files");
-            Console.WriteLine("Dotnet: upload success as " +  upload_response);
+            client.info("Dotnet: upload success as " +  upload_response);
 
             var eventcount = 0;
             var watch_response = await client.watch("entities", "", (eventObj) => {
-                Console.WriteLine("watch event " + eventObj.operation + " on " + eventObj.document);
+                client.info("watch event " + eventObj.operation + " on " + eventObj.document);
                 eventcount++;
             });
-            Console.WriteLine("Dotnet: watch registered success as " +  watch_response);
+            client.info("Dotnet: watch registered success as " +  watch_response);
 
             var insert_many_result = await client.InsertMany("entities", "[{\"name\": \"test from dotnet 1 \", \"_type\": \"test\"}, {\"name\": \"test from dotnet 2\", \"_type\": \"test\"}]");
 
@@ -161,10 +166,10 @@ class Program
 
             var queuecount = 0;
             var register_queue_response = await client.RegisterQueue("test2queue", (eventObj) => {
-                Console.WriteLine("watch event " + eventObj.queuename + " on " + eventObj.data);
+                client.info("watch event " + eventObj.queuename + " on " + eventObj.data);
                 queuecount++;
             });
-            Console.WriteLine("Dotnet: registered queue success as " + register_queue_response);
+            client.info("Dotnet: registered queue success as " + register_queue_response);
 
             await client.QueueMessage("{\"name\": \"test message 1 \"}", "test2queue");
             await client.QueueMessage("{\"name\": \"test message 2 \"}", "test2queue");
@@ -177,10 +182,10 @@ class Program
 
             var exchangecount = 0;
             var register_exchange_response = await client.RegisterExchange("testexc", eventHandler: (eventObj) => {
-                Console.WriteLine("watch event " + eventObj.queuename + " on " + eventObj.data);
+                client.info("watch event " + eventObj.queuename + " on " + eventObj.data);
                 exchangecount++;
             });
-            Console.WriteLine("Dotnet: registered exchange success, using queue " + register_exchange_response);
+            client.info("Dotnet: registered exchange success, using queue " + register_exchange_response);
 
             await client.QueueMessage("{\"name\": \"test message 1 \"}", exchangename: "testexc");
             await client.QueueMessage("{\"name\": \"test message 2 \"}", exchangename: "testexc");
@@ -193,10 +198,10 @@ class Program
 
 
             var count_response = await client.Count("entities", "");
-            Console.WriteLine("Dotnet: count success as " +  count_response);
+            client.info("Dotnet: count success as " +  count_response);
 
             var distinct_response = await client.Distinct("entities", "_type");
-            Console.WriteLine("Dotnet: distinct success as " + string.Join(",", distinct_response));
+            client.info("Dotnet: distinct success as " + string.Join(",", distinct_response));
 
             // while(true) {
             //     await Task.Delay(1000);
